Updates to support most recent Ember Data conventions
More recent Ember Data requires all JSON keys to use camel
casing instead of underscores. In addition, the outer key
with name of relation needs to be pluralized when it contains
a list of model instances.

This makes use of the Python inflection library which can convert
between both camel case and underscore formatting, as well as singular
and plural tenses of english words.