ResourceRelatedField validation
Previously, if a ResourceRelatedField was passed as input an invalid
resource identifier object missing either the 'type' or 'id' attributes,
this resulted in an uncaught KeyError and a 500 error. This commit
adds validation that the object contains both a 'type' and 'id' attribute.
It fails with a helpful error message.