Rename comment_set to comments, use json viewset
Updates the tests to use the optimised queryset views provided in
views.py. For these to work, the serializer fields need to match the
model names so I have just changed the related_name in the test models.

This could be improved elsewhere, with the queryset prefetching based on
the serializer. But for the most basic use case, this is fine.