Fix for apps that don't load contenttypes module.
PR #319 brought support for generic relations. Unfortunately apps that
don't add contenttypes to it's INSTALLED_APPS would crash and burn:

```
RuntimeError: Model class django.contrib.contenttypes.models.ContentType doesn't declare an explicit app_label and isn't in an application in INSTALLED_APPS.
```

Also using `type(something) is object()` comparison as a safer
alternative to `type(something) is type(None)`. If `something` happened
to be `None` we would enter a branch that was never supposed to run.