Override the init in setUp and tearDown
The previous approach to restore the overridden init was not running if the
test failed, so it was only restoring the init when the test passed. By setting
it and restoring it during the setUp and tearDown of the test we ensure that it
will always be restored.

Also because the overridden init is now declared in the setUp, I have moved the
asserts back to the test for readability.