Pass on instance when using polymorphic serializers (#764)
Fixes #759 

Pass self.instance as the first parameter when initializing the child serializer from a polymorphic serialiser. This does not affect basic functionality, but makes the child serializer instance consistent with its usage of instance and data, which can fix potential issues when extending serilaizers that are part of a polymorphic serializer.