Add example code to make the test-suite fail
When adding a related resource to an inherited polymorphic model, that
is absent on the base model, the related resource type is incorrectly
resolved in get_related_resource_type.

Introduce a LabResults model to the example code and add it as a related
resource on the ResearchProjectSerializer to provoke an error. This
already causes the test-suite to fail, so no new test has been added.