Verified in all fields whether reserved field names are used
In meta class only specifically declared fields in serializer are accessible.
In case of `ModelSerializer` fields may be just on the model or
if a user has overwritten what fields are being returned during runtime
DJA won't notice. So there does not seem to be another way then simply hooking
into the `get_fields` method to check for reserved field names.