Deprecated usage of `type` field name
This support was added in #376 but only for non polymorphic fields.

However as per specification [0] `type` must not be a field name and
therefore must be forbidden in DJA as well.

Some dependents might depend on being allowed to have a field
name `type` so deprecating it now and remove it in next major version.

[0] https://jsonapi.org/format/#document-resource-object-fields